pyproject.toml
[project]
name = "openai-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.4.2",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "datetime>=5.5",
]
